+++ GreenWallet/GreenWallet/ViewControllers/ImportMnemonicViewController.swift
@@ -169,7 +169,7 @@ class ImportMnemonicViewController: UIViewController {
         self.present(spinnerVC, animated: true)
         
         var adreses = ""
-        var balances: [Int] = []
+        var balances: [Double] = []
         var walletsDict: [Int] = []
         var privateKey = ChiaPrivate(private_key: ChiaPrivateKey(farmer_pk: "", fingerprint: 0, pk: "", pool_pk: "", seed: "", sk: ""), success: true)
         var name = ""
